Update dejure-plugin to version 1.6 from upstream. (#58)
* Update dejure-plugin to version 1.6 from upstream.

v1.6 was not yet published.

Improve fsockopen, switch to SSL, add user agent.

Signed-off-by: Thomas Hochstein <thh@inter.net>

* Bugfix: Don't cache results after errors.

djo_vernetzen_ueber_dejure_org() will return FALSE
instead of an output text if the API call fails
or results in any error.

But the "master" function djo_vernetzen() will call
integritaetskontrolle() anyway. The latter function
checks whether input and output texts match after
removing all added links. If the check succeeds, it
will return the output text; if the check fails, the
input text. So the result will always be TRUE.

And only after this will cache the output text if
it's not FALSE - which it can't be, ever. So if the
API call fails, the unmodified input text will be
cached.

But if we don't call integritaetskontrolle() after
API failures - which isn't necessary as the output
text is just FALSE anyway -, the FALSE will carry
through so nothing is cached.

Signed-off-by: Thomas Hochstein <thh@inter.net>