Reset 404 error message for external PHP plugin (#79)
Just like the static pages plugin the external PHP plugin overrides the 404 error handling. The content from the external PHP script is displayed as expected but above this content the error message from the language constant URL_NOT_FOUND is shown. The static pages plugin resets this error message to get rid of the error message. The change in this commit does exactly the same thing for the external PHP plugin.