Get language selector working.
The chosen language got lost (reset to "en") after following the next
link, as the language field wasn't part of those links.

Now we choose the language in three steps:

If "language" ist set (after using the language selector form), we use
that.

If it isn't set but a plugin group was chosen, we take the right
language from the filename of that page.

Only if neither is set we fall back to english.

Moreover we keep the current language "selected" in the language
selector dropdown.

Signed-off-by: Thomas Hochstein <thh@inter.net>