[event_social] Set social media image for entry.
evenmt_social will use the first image from the
entry for the OpenGrahp "og:image" meta property
used by Facebook and Twitter; if there is none,
it will fall back to the default image from the
plugin configuration.

The timeline and photo themes support setting
a featured image that is used instead.

Add the possibility to set a social media
image for each entry, regardless of theme.
If there is none set, fall back to timeline
or photo featured images, or the first image
in the enttry body, or the default image from
the plugin configuration (in that order).

Signed-off-by: Thomas Hochstein <thh@inter.net>