bayes: make tokens case-sensitive on MySQL (#161)
Alter the token column in the b8_wordlist to a case-sensitive
collation to prevent duplicate key errors when the same token is
encountered in different cases (e.g. 'the', 'THE' and 'ThE').

The PHP code compares tokens case-sensitive when deciding whether to
do an INSERT or UPDATE, so the database also needs to be
case-sensitive.